requirements.txt
pylint == 2.15.3
black == 22.3.0
pytest == 7.1.3
ruff == 0.0.284
fire == 0.7.0
requests == 2.32.3
pandas == 2.2.2
python-dotenv == 1.0.1
databricks-sql-connector == 3.4.0
pyspark.sql
subprocess